refactored  [hubUrl]/browse.tsx into smaller components as preparation for subhub page
