+++ frontend/pages/hubs/[hubUrl]/browse.tsx
@@ -1,10 +1,9 @@
-import { Fab, Typography, useMediaQuery } from "@mui/material";
+import { useMediaQuery } from "@mui/material";
 import makeStyles from "@mui/styles/makeStyles";
 import parseHtml from "html-react-parser";
 import Head from "next/head";
-import React, { useContext, useEffect, useRef, useState } from "react";
+import { useContext, useEffect, useRef, useState } from "react";
 import Cookies from "universal-cookie";
-import { apiRequest, getLocalePrefix } from "../../../public/lib/apiOperations";
 import {
   getOrganizationTagsOptions,
   getProjectTagsOptions,
@@ -18,15 +17,11 @@ import getTexts from "../../../public/texts/texts";
 import BrowseContent from "../../../src/components/browse/BrowseContent";
 import UserContext from "../../../src/components/context/UserContext";
 import BrowseExplainer from "../../../src/components/hub/BrowseExplainer";
-import FashionDescription from "../../../src/components/hub/description/FashionDescription";
-import FoodDescription from "../../../src/components/hub/description/FoodDescription";
 import HubContent from "../../../src/components/hub/HubContent";
 import HubHeaderImage from "../../../src/components/hub/HubHeaderImage";
 import NavigationSubHeader from "../../../src/components/hub/NavigationSubHeader";
 import WideLayout from "../../../src/components/layouts/WideLayout";
 import DonationCampaignInformation from "../../../src/components/staticpages/donate/DonationCampaignInformation";
-import { retrievePage } from "../../../src/utils/webflow";
-import AddIcon from "@mui/icons-material/Add";
 import { Theme } from "@mui/material/styles";
 import theme from "../../../src/themes/hubTheme";
 import BrowseContext from "../../../src/components/context/BrowseContext";
@@ -34,53 +29,20 @@ import { transformThemeData } from "../../../src/themes/transformThemeData";
 import getHubTheme from "../../../src/themes/fetchHubTheme";
 import isLocationHubLikeHub from "../../../public/lib/isLocationHubLikeHub";
 import { FilterProvider } from "../../../src/components/provider/FilterProvider";
+import {
+  getHubAmbassadorData,
+  getHubData,
+  getHubSupportersData,
+} from "../../../public/lib/getHubData";
+import { retrieveDescriptionFromWebflow } from "../../../src/utils/webflow";
+import { HubDescription } from "../../../src/components/hub/description/HubDescription";
+import { FabShareButton } from "../../../src/components/hub/FabShareButton";
 
 const useStyles = makeStyles((theme) => ({
-  moreInfoSoon: {
-    fontWeight: 600,
-    maxWidth: 800,
-    marginTop: theme.spacing(2),
-    textAlign: "center",
-  },
   content: {
     position: "relative",
   },
 }));
-
-type ShareProjectMakeStyleProps = {
-  isCustomHub: boolean;
-};
-
-const shareProjectFabStyle = makeStyles((theme) => ({
-  fabShareProject: (props: ShareProjectMakeStyleProps) => ({
-    position: "fixed",
-    background: props.isCustomHub
-      ? theme.palette.background.default_contrastText
-      : theme.palette.primary.light,
-    color: props.isCustomHub ? theme.palette.background.default : "default",
-    // bottom: theme.spacing(5),
-    right: theme.spacing(3),
-  }),
-}));
-
-const DESCRIPTION_WEBFLOW_LINKS = {
-  energy: {
-    en: "energy-en",
-    de: "energie-de",
-  },
-  mobility: {
-    de: "mobilitat-de",
-    en: "mobility-en",
-  },
-  biodiversity: {
-    de: "biodiversitat",
-    en: "biodiversity-en",
-  },
-  landuse: {
-    de: "landuse-de",
-    en: "landuse-en",
-  },
-};
 
 //potentially switch back to getinitialprops here?!
 export async function getServerSideProps(ctx) {
@@ -322,98 +284,3 @@ export default function Hub({
     </>
   );
 }
-
-const FabShareButton = ({ locale, hubAmbassador, isCustomHub, hubUrl }) => {
-  const fabClass = shareProjectFabStyle({ isCustomHub: isCustomHub });
-  const queryString = hubUrl ? `?hub=${hubUrl}` : "";
-  return (
-    <Fab
-      className={fabClass.fabShareProject}
-      size="medium"
-      color="primary"
-      href={`${getLocalePrefix(locale)}/share${queryString}`}
-      sx={{ bottom: (theme) => (hubAmbassador ? theme.spacing(11.5) : theme.spacing(5)) }}
-      // onClick={}
-    >
-      <AddIcon />
-    </Fab>
-  );
-};
-
-const HubDescription = ({ hub, texts }) => {
-  const classes = useStyles();
-  if (hub === "food") return <FoodDescription />;
-  if (hub === "fashion") return <FashionDescription />;
-  return (
-    <Typography className={classes.moreInfoSoon}>
-      {texts.more_info_about_hub_coming_soon}
-    </Typography>
-  );
-};
-
-const WEBFLOW_BASE_LINK = "https://climateconnect.webflow.io/hub-texts/";
-
-const retrieveDescriptionFromWebflow = async (query, locale) => {
-  if (
-    DESCRIPTION_WEBFLOW_LINKS[query?.hubUrl] &&
-    DESCRIPTION_WEBFLOW_LINKS[query?.hubUrl][locale]
-  ) {
-    const props = await retrievePage(
-      WEBFLOW_BASE_LINK + DESCRIPTION_WEBFLOW_LINKS[query.hubUrl][locale]
-    );
-    return props;
-  }
-  return null;
-};
-
-const getHubData = async (url_slug, locale) => {
-  try {
-    const resp = await apiRequest({
-      method: "get",
-      url: `/api/hubs/${url_slug}/`,
-      locale: locale,
-    });
-    return resp.data;
-  } catch (err: any) {
-    if (err.response && err.response.data) {
-      console.log(err.response.data);
-      console.error("Error in getHubData!: " + err.response.data.detail);
-    }
-    return null;
-  }
-};
-
-const getHubAmbassadorData = async (url_slug, locale) => {
-  try {
-    const resp = await apiRequest({
-      method: "get",
-      url: `/api/hubs/${url_slug}/ambassador/`,
-      locale: locale,
-    });
-    return resp.data;
-  } catch (err: any) {
-    if (err.response && err.response.data)
-      console.log("Error in getHubAmbassadorData: " + err.response.data.detail);
-    console.log(err);
-    return null;
-  }
-};
-const getHubSupportersData = async (url_slug, locale) => {
-  try {
-    const resp = await apiRequest({
-      method: "get",
-      url: `/api/hubs/${url_slug}/supporters/`,
-      locale: locale,
-    });
-    return resp.data;
-  } catch (err: any) {
-    //Don't log an error if there simply are no supporters for this hub
-    if (err?.response?.status === 404) {
-      return null;
-    }
-    if (err.response && err.response.data)
-      console.log("Error in getHubSupportersData: " + err.response.data.detail);
-    console.log(err);
-    return null;
-  }
-};
